Added htmx to currency converter + added api
