+++ backend/api/currency_converter/urls.py
@@ -1,0 +1,13 @@
+from django.urls import path
+
+from . import convert
+
+urlpatterns = [
+    path(
+        "convert/",
+        convert.currency_conversion,
+        name="convert",
+    ),
+]
+
+app_name = "currency_converter"
