+++ frontend/templates/pages/currency_converter/dashboard.html
@@ -3,7 +3,9 @@
     <p class="text-white font-bold">{{ page }}</p>
     {# TODO SEARCH BAR #}
 
-    <form method="post" class="my-4 max-w-screen-xl card border border-gray-700 dark:border-gray-700">
+    <form class="my-4 max-w-screen-xl card border border-gray-700 dark:border-gray-700"
+          hx-post="{% url 'api:currency_converter:convert' %}" hx-target='div[data-type="results"]'
+          method="post">
         {% csrf_token %}
         <div class="mb-4 flex w-full flex-col">
 
@@ -12,7 +14,7 @@
                 <div class="form-field">
                     <label class="text-lg font-semibold text-gray-900 dark:text-white mb-1">Currency Amount</label>
                     <br>
-                    <input name="currency_amount" placeholder="100" type="text"
+                    <input name="currency_amount" placeholder="100" max="9999999999" maxlength="10" type="text"
                            class="peer input-ghost-secondary input input-bordered input-block">
                     <label class="form-label peer-[&amp;:not(:placeholder-shown):not(:focus):invalid]:block hidden ">
                         <span class="form-label-alt">Please enter a valid currency amount name.</span>
@@ -46,16 +48,11 @@
                     </div>
                 </div>
 
-                {% if converted_amount %}
-                    <div class="text-lg text-center text-gray-900 dark:text-white">
-                        <p>
-                            <strong>{{ original_currency_sign }}{{ original_amount }}</strong> in {{ original_currency }} is
-                            <strong>{{ target_currency_sign }}{{ converted_amount }}</strong> in {{ target_currency }}
-                        </p>
-                    </div>
-                {% endif %}
+                <div data-type="results"></div>
 
-                <button class="btn btn-primary mb-2">Convert</button>
+                <button class="btn btn-primary mb-2 loading-htmx" data-type="convert-btn">
+                    Convert
+                </button>
             </div>
