+++ frontend/templates/partials/messages_list.html
@@ -1,0 +1,1 @@
+{% component "messages_list" %}
